style(workspace): unstyled title input, reading-source row in bulk drafts
- Pass `unstyled` to the title Input in both the single-column and
  bulk-add dialogs. The Input primitive's own bordered chrome (and its
  ::before inner shadow) painted a visible second ring inside the
  outer card frame; `unstyled` removes it cleanly.
- Lift ReadingFromRow + ReadingChip + FileChip from create-property.tsx
  into composer-primitives.tsx so bulk drafts can render the same
  "Čte z [Documents]" affordance.
- Seed each bulk draft's file dependencies with the workspace's file
  properties by default and combine fileIds + @-mentions into the
  batch payload's dependencies array on submit.
